+++ sandpack-react/src/components/CodeEditor/utils.ts
@@ -1,12 +1,12 @@
-import { HighlightStyle, tags } from "@codemirror/highlight";
 import { css } from "@codemirror/lang-css";
 import { html } from "@codemirror/lang-html";
 import { javascript } from "@codemirror/lang-javascript";
 import type { LanguageSupport } from "@codemirror/language";
-import type { Extension } from "@codemirror/state";
-import type { Text } from "@codemirror/text";
+import { HighlightStyle } from "@codemirror/language";
+import type { Extension, Text } from "@codemirror/state";
 import { EditorView } from "@codemirror/view";
 import * as React from "react";
+import { tags } from "@lezer/highlight";
 
 import { THEME_PREFIX } from "../../styles";
 import type { CustomLanguage, SandpackTheme } from "../../types";
@@ -119,18 +119,25 @@ export const getSyntaxHighlight = (theme: SandpackTheme): HighlightStyle =>
       class: classNameToken("static"),
     },
     {
-      tag: tags.tagName,
+      tag: tags.variableName,
+      class: classNameToken("plain"),
+    },
+    {
+      // Standard tags, e.g <h1 />
+      tag: tags.standard(tags.tagName),
       class: classNameToken("tag"),
     },
-    { tag: tags.variableName, class: classNameToken("plain") },
     {
-      // Highlight function call
-      tag: tags.function(tags.variableName),
-      class: classNameToken("definition"),
-    },
-    {
-      // Highlight function definition differently (eg: functional component def in React)
-      tag: tags.definition(tags.function(tags.variableName)),
+      tag: [
+        // Highlight function call
+        tags.function(tags.variableName),
+
+        // Highlight function definition differently (eg: functional component def in React)
+        tags.definition(tags.function(tags.variableName)),
+
+        // "Custom tags", meaning React component
+        tags.tagName,
+      ],
       class: classNameToken("definition"),
     },
     {