+++ sandpack-react/src/components/common/Layout.tsx
@@ -5,8 +5,7 @@ import { useSandpack } from "../../hooks/useSandpack";
 import { css, THEME_PREFIX } from "../../styles";
 import { classNames } from "../../utils/classNames";
 import { useCombinedRefs } from "../CodeEditor/utils";
-
-import { stackClassName } from ".";
+import { stackClassName } from "./Stack";
 
 /**
  * @category Components
@@ -33,7 +32,6 @@ export const layoutClassName = css({
     flexGrow: 1,
     flexShrink: 1,
     flexBasis: "0",
-    minWidth: "350px",
     height: "$layout$height",
 
     "@media print": {
@@ -42,10 +40,13 @@ export const layoutClassName = css({
     },
 
     "@media screen and (max-width: 768px)": {
-      height: "auto",
+      [`&:not(.${THEME_PREFIX}-preview, .${THEME_PREFIX}-editor, .${THEME_PREFIX}-preset-column)`]:
+        {
+          height: "calc($layout$height / 2)",
+        },
 
       /* triggers the layout break at the 768px breakpoint, not when the component is less then 700px */
-      minWidth: "100% !important;",
+      minWidth: "100%",
     },
   },