+++ sandpack-react/src/templates/index.tsx
@@ -8,6 +8,7 @@ import { VANILLA_TEMPLATE } from "./vanilla";
 import { VANILLA_TYPESCRIPT_TEMPLATE } from "./vanilla-typescript";
 import { VUE_TEMPLATE } from "./vue";
 import { VUE_TEMPLATE_3 } from "./vue3";
+import { VUE_TS_TEMPLATE_3 } from "./vue3-ts";
 export { ANGULAR_TEMPLATE } from "./angular";
 export { REACT_TEMPLATE } from "./react";
 export { REACT_TYPESCRIPT_TEMPLATE } from "./react-typescript";
@@ -18,6 +19,7 @@ export { VANILLA_TEMPLATE } from "./vanilla";
 export { VANILLA_TYPESCRIPT_TEMPLATE } from "./vanilla-typescript";
 export { VUE_TEMPLATE } from "./vue";
 export { VUE_TEMPLATE_3 } from "./vue3";
+export { VUE_TS_TEMPLATE_3 } from "./vue3-ts";
 
 /**
  * @hidden
@@ -29,6 +31,7 @@ export const SANDBOX_TEMPLATES = {
   vanilla: VANILLA_TEMPLATE,
   "vanilla-ts": VANILLA_TYPESCRIPT_TEMPLATE,
   vue3: VUE_TEMPLATE_3,
+  "vue3-ts": VUE_TS_TEMPLATE_3,
   angular: ANGULAR_TEMPLATE,
   svelte: SVELTE_TEMPLATE,
   solid: SOLID_TEMPLATE,